More WASM work
Former-commit-id: 100877b4c1d64f72f770e52bd4c69afa1f739a17
Former-commit-id: 681a7ba4c4beea305d18e5eec73aa352e7cae0e6 [formerly 8629acda07bba018e1684c79f32ef383b17dde79]
Former-commit-id: e6638f60e093c2d33e083ecf6d793622f607a481

+++ modules/renderer/source/hip/hiprenderer/backend/gl/glrenderer.d
@@ -153,7 +153,7 @@ class Hip_GL3Renderer : IHipRendererImpl
         else
             return new Hip_GL_VertexArrayObject();
     }
-    public IHipVertexBufferImpl createVertexBuffer(ulong size, HipBufferUsage usage)
+    public IHipVertexBufferImpl createVertexBuffer(size_t size, HipBufferUsage usage)
     {
         return new Hip_GL3_VertexBufferObject(size, usage);
     }
